requirements.txt
amqp==2.3.2
anyjson==0.3.3
Babel==2.6.0
backports-abc==0.5
-e git+https://github.com/celery/billiard.git@af851466df6f083dea36bbe21a40862655ec0339#egg=billiard
celery==4.2.1
dj-static==0.0.6
Django==1.11.15
django-celery==3.2.2
djangorestframework==3.8.2
flower==0.9.2
futures==3.2.0
gunicorn==19.9.0
-e git+https://github.com/celery/kombu.git@ddb0c8ffadf2822480e2cbacf0c3ebc3167d9536#egg=kombu
paho-mqtt==1.3.1
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
redis==2.10.6
singledispatch==3.4.0.3
six==1.11.0
static3==0.7.0
tornado==5.1
vine==1.1.4
